Workaround docusaurus issue with pagination
When a document id is on the sidebar twice consecutively, the pagination will always calculate you are on the first one. The pagination will then see the next one, and assume that is the next document you should paginate to. When you click the button, you are sent back to the first document (from the perspective of the sidebar).